+++ st2reactor/st2reactor/cmd/sensormanager.py
@@ -1,16 +1,30 @@
+# Licensed to the StackStorm, Inc ('StackStorm') under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os
 import sys
 
 import eventlet
-from oslo_config import cfg
 
 from st2common import log as logging
 from st2common.service_setup import setup as common_setup
 from st2common.service_setup import teardown as common_teardown
 from st2common.exceptions.sensors import SensorNotFoundException
-from st2common.persistence.sensor import SensorType
 from st2reactor.sensor import config
 from st2reactor.container.manager import SensorContainerManager
+from st2reactor.container.partitioner_lookup import get_sensors_partitioner
 
 eventlet.monkey_patch(
     os=True,
@@ -32,36 +46,12 @@ def _teardown():
     common_teardown()
 
 
-def _get_all_enabled_sensors():
-    # only query for enabled sensors.
-    sensors = SensorType.query(enabled=True)
-    LOG.info('Found %d registered sensors in db scan.', len(sensors))
-    return sensors
-
-
 def main():
     try:
         _setup()
-        container_manager = SensorContainerManager()
-        sensors = None
-
-        if cfg.CONF.sensor_ref:
-            # pick the sensor even if it is disabled since it has been specifically
-            # asked to be run. This is debug/test usecase so it is reasonable to run
-            # a disabled sensor.
-            sensor = SensorType.get_by_ref(cfg.CONF.sensor_ref)
-            if not sensor:
-                raise SensorNotFoundException('Sensor %s not found in db.' % cfg.CONF.sensor_ref)
-            sensors = [sensor]
-        else:
-            sensors = _get_all_enabled_sensors()
-
-        if not sensors:
-            msg = 'No sensors configured to run. See http://docs.stackstorm.com/sensors.html. ' + \
-                  'Register some sensors and a container will pick them to run.'
-            LOG.info(msg)
-
-        return container_manager.run_sensors(sensors=sensors)
+        sensors_partitioner = get_sensors_partitioner()
+        container_manager = SensorContainerManager(sensors_partitioner=sensors_partitioner)
+        return container_manager.run_sensors()
     except SystemExit as exit_code:
         return exit_code
     except SensorNotFoundException as e: