+++ st2reactor/st2reactor/container/manager.py
@@ -28,23 +28,26 @@ LOG = logging.getLogger(__name__)
 
 
 class SensorContainerManager(object):
-    # TODO: Load balancing for sensors.
-    def __init__(self, max_containers=10):
-        self._max_containers = max_containers
+
+    def __init__(self, sensors_partitioner):
         self._sensor_container = None
         self._sensors_watcher = SensorWatcher(create_handler=self._handle_create_sensor,
                                               update_handler=self._handle_update_sensor,
                                               delete_handler=self._handle_delete_sensor,
                                               queue_suffix='sensor_container')
         self._container_thread = None
+        if not sensors_partitioner:
+            raise ValueError('sensors_partitioner should be non-None.')
+        self._sensors_partitioner = sensors_partitioner
 
-    def run_sensors(self, sensors):
+    def run_sensors(self):
         """
-        :param sensors: A list of DB models of sensors to run.
-        :type sensors: ``list``
+        Run all sensors as determined by sensors_partitioner.
         """
+        sensors = self._sensors_partitioner.get_sensors()
         if sensors:
             LOG.info('Setting up container to run %d sensors.', len(sensors))
+            LOG.info('\tSensors list - %s.', [self._get_sensor_ref(sensor) for sensor in sensors])
 
         sensors_to_run = []
         for sensor in sensors:
@@ -71,7 +74,8 @@ class SensorContainerManager(object):
             LOG.info('(PID:%s) SensorContainer stopped. Reason - %s', os.getpid(),
                      sys.exc_info()[0].__name__)
 
-            self._container_thread = eventlet.kill(self._container_thread)
+            eventlet.kill(self._container_thread)
+            self._container_thread = None
 
             return 0
 
@@ -106,10 +110,19 @@ class SensorContainerManager(object):
     #################################################
 
     def _handle_create_sensor(self, sensor):
+        if not self._sensors_partitioner.is_sensor_owner(sensor):
+            LOG.info('sensor %s is not supported. Ignoring create.', self._get_sensor_ref(sensor))
+            return
+        if not sensor.enabled:
+            LOG.info('sensor %s is not enabled.', self._get_sensor_ref(sensor))
+            return
         LOG.info('Adding sensor %s.', self._get_sensor_ref(sensor))
         self._sensor_container.add_sensor(sensor=self._to_sensor_object(sensor))
 
     def _handle_update_sensor(self, sensor):
+        if not self._sensors_partitioner.is_sensor_owner(sensor):
+            LOG.info('sensor %s is not supported. Ignoring update.', self._get_sensor_ref(sensor))
+            return
         sensor_ref = self._get_sensor_ref(sensor)
         LOG.info('Sensor %s updated. Reloading sensor.', sensor_ref)
         sensor_obj = self._to_sensor_object(sensor)
@@ -122,6 +135,9 @@ class SensorContainerManager(object):
             LOG.info('Sensor %s reloaded.', sensor_ref)
 
     def _handle_delete_sensor(self, sensor):
+        if not self._sensors_partitioner.is_sensor_owner(sensor):
+            LOG.info('sensor %s is not supported. Ignoring delete.', self._get_sensor_ref(sensor))
+            return
         LOG.info('Unloading sensor %s.', self._get_sensor_ref(sensor))
         self._sensor_container.remove_sensor(sensor=self._to_sensor_object(sensor))