+++ st2reactor/st2reactor/sensor/config.py
@@ -13,9 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from oslo_config import cfg
+from oslo_config import cfg, types
 
 from st2common import config as st2cfg
+from st2common.constants.sensors import DEFAULT_PARTITION_LOADER
 from st2common.constants.system import VERSION_STRING
 
 CONF = cfg.CONF
@@ -44,6 +45,15 @@ def _register_sensor_container_opts(ignore_errors=False):
                    help='location of the logging.conf file')
     ]
     st2cfg.do_register_opts(logging_opts, group='sensorcontainer', ignore_errors=ignore_errors)
+
+    partition_opts = [
+        cfg.StrOpt('sensor_node_name', default='sensornode1',
+                   help='name of the sensor node.'),
+        cfg.Opt('partition_provider', type=types.Dict(value_type=types.String()),
+                default={'name': DEFAULT_PARTITION_LOADER},
+                help='Provider of sensor node partition config.')
+    ]
+    st2cfg.do_register_opts(partition_opts, group='sensorcontainer', ignore_errors=ignore_errors)
 
     sensor_test_opt = cfg.StrOpt('sensor-ref', help='Only run sensor with the provided reference. \
         Value is of the form pack.sensor-name.')